fix(dev): inline transitive workflow imports in generation bundles
Third-party dependencies inlined into a generation bundle can import
`@workflow/*` (e.g. `@ai-sdk/provider-utils` -> `@workflow/serde`). eve
vendors those packages, so leaving the import external emits a bare
specifier no materialized generation can resolve. Externalize workflow
runtime imports only for authored importers and inline them inside
bundled node_modules code.

Also exclude the channel identity plugin from generation bundles so an
immutable artifact structurally cannot read the process-global channel
cache, and split the package boundary plugins into their own module.
